extract pages
